Refactor ASM code into shared module, fix architecture filtering
- Create asm_common.py with shared syscall/utility functions
- Remove duplicated code from assembler.py and nasmgen.py
- Fix label vs directive classification bug (local labels like .Lloop:)
- Add x86_64 variants for all string.plib functions
- Add architecture filtering to prevent duplicate symbol errors
- Fix plstd/string.plib: add space declaration, error checking in strdup
